feat: optimistic lock exception handling
