+++ src/main/java/io/github/onecx/user/profile/rs/external/v1/controllers/UserProfileV1RestController.java
@@ -3,30 +3,20 @@ package io.github.onecx.user.profile.rs.external.v1.controllers;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
-import jakarta.validation.ConstraintViolationException;
 import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.core.UriInfo;
 
-import org.apache.commons.lang3.StringUtils;
-import org.jboss.resteasy.reactive.RestResponse;
-import org.jboss.resteasy.reactive.server.ServerExceptionMapper;
 import org.tkit.quarkus.context.ApplicationContext;
-import org.tkit.quarkus.jpa.exceptions.ConstraintException;
 import org.tkit.quarkus.log.cdi.LogService;
 
 import gen.io.github.onecx.user.profile.rs.external.v1.UserProfileV1Api;
-import gen.io.github.onecx.user.profile.rs.external.v1.model.CreateUserPreferenceDTO;
-import gen.io.github.onecx.user.profile.rs.external.v1.model.ProblemDetailResponseDTO;
-import gen.io.github.onecx.user.profile.rs.external.v1.model.UpdateUserPersonDTO;
-import gen.io.github.onecx.user.profile.rs.external.v1.model.UpdateUserSettingsDTO;
 import io.github.onecx.user.profile.domain.daos.PreferenceDAO;
 import io.github.onecx.user.profile.domain.daos.UserProfileDAO;
 import io.github.onecx.user.profile.domain.models.UserProfile;
-import io.github.onecx.user.profile.rs.external.v1.mappers.PreferenceMapper;
-import io.github.onecx.user.profile.rs.external.v1.mappers.UserProfileMapper;
-import io.github.onecx.user.profile.rs.external.v1.mappers.V1ExceptionMapper;
-import io.github.onecx.user.profile.rs.external.v1.service.JWTService;
+import io.github.onecx.user.profile.rs.external.v1.mappers.PreferenceV1Mapper;
+import io.github.onecx.user.profile.rs.external.v1.mappers.UserProfileV1Mapper;
+import io.github.onecx.user.profile.rs.internal.service.JWTService;
 
 @LogService
 @ApplicationScoped
@@ -40,76 +30,16 @@ public class UserProfileV1RestController implements UserProfileV1Api {
     UserProfileDAO userProfileDAO;
 
     @Inject
-    V1ExceptionMapper exceptionMapper;
+    PreferenceV1Mapper preferenceMapper;
 
     @Inject
-    PreferenceMapper preferenceMapper;
-
-    @Inject
-    UserProfileMapper userProfileMapper;
+    UserProfileV1Mapper userProfileV1Mapper;
 
     @Inject
     JWTService jwtService;
 
     @Context
     UriInfo uriInfo;
-
-    @Override
-    public Response createUserPreference(CreateUserPreferenceDTO createUserPreferenceDTO) {
-        var userId = ApplicationContext.get().getPrincipal();
-        var userProfile = userProfileDAO.getUserProfileByUserId(userId, UserProfile.ENTITY_GRAPH_LOAD_ALL);
-
-        if (userProfile == null) {
-            throw new ConstraintException("Profile for userID: " + userId + " does not exist",
-                    V1ErrorKeys.USER_PROFILE_DOES_NOT_EXIST, null);
-        }
-        var gotPreference = preferenceMapper.create(createUserPreferenceDTO);
-        gotPreference.setUserProfile(userProfile);
-
-        var createdPreference = preferenceDAO.create(gotPreference);
-        var createdPreferenceDTOv1 = preferenceMapper.map(createdPreference);
-
-        return Response
-                .created(uriInfo.getAbsolutePathBuilder().path(createdPreference.getId()).build())
-                .entity(createdPreferenceDTOv1)
-                .build();
-    }
-
-    @Override
-    @Transactional
-    public Response deleteUserPreference(String id) {
-        var userId = ApplicationContext.get().getPrincipal();
-
-        var preference = preferenceDAO.findById(id);
-
-        if (preference != null) {
-            if (!preference.getUserProfile().getUserId().equals(userId)) {
-                throw new ConstraintException("Preference is not from actual user", V1ErrorKeys.PREFERENCE_NOT_FROM_ACTUAL_USER,
-                        null);
-            } else {
-                preferenceDAO.delete(preference);
-            }
-        }
-
-        return Response.noContent().build();
-    }
-
-    @Override
-    @Transactional
-    public Response deleteUserProfile() {
-        var userId = ApplicationContext.get().getPrincipal();
-
-        var userProfile = userProfileDAO.getUserProfileByUserId(userId, UserProfile.ENTITY_GRAPH_LOAD_ALL);
-
-        if (userProfile != null) {
-            var preferences = preferenceDAO.getAllPreferencesByUserId(userId);
-            preferenceDAO.delete(preferences);
-            userProfileDAO.delete(userProfile);
-
-        }
-
-        return Response.noContent().build();
-    }
 
     @Override
     public Response getUserPerson() {
@@ -121,7 +51,7 @@ public class UserProfileV1RestController implements UserProfileV1Api {
             return Response.status(Response.Status.NOT_FOUND).build();
         }
 
-        var userPerson = userProfileMapper.map(userProfile.getPerson());
+        var userPerson = userProfileV1Mapper.map(userProfile.getPerson());
 
         return Response.ok(userPerson).build();
     }
@@ -147,7 +77,7 @@ public class UserProfileV1RestController implements UserProfileV1Api {
             userProfile = userProfileDAO.create(createUserProfile);
         }
 
-        return Response.ok(userProfileMapper.map(userProfile)).build();
+        return Response.ok(userProfileV1Mapper.map(userProfile)).build();
     }
 
     @Override
@@ -160,78 +90,7 @@ public class UserProfileV1RestController implements UserProfileV1Api {
             return Response.status(Response.Status.NOT_FOUND).build();
         }
 
-        return Response.ok(userProfileMapper.map(userProfile.getAccountSettings())).build();
+        return Response.ok(userProfileV1Mapper.map(userProfile.getAccountSettings())).build();
     }
 
-    @Override
-    public Response updateUserPerson(UpdateUserPersonDTO updateUserPersonDTO) {
-        var userId = ApplicationContext.get().getPrincipal();
-
-        var userProfile = userProfileDAO.getUserProfileByUserId(userId, UserProfile.ENTITY_GRAPH_LOAD_PERSON);
-        if (userProfile == null) {
-            return Response.status(Response.Status.NOT_FOUND).build();
-        }
-
-        userProfileMapper.update(userProfile.getPerson(), updateUserPersonDTO);
-        userProfile = userProfileDAO.update(userProfile);
-
-        return Response.ok(userProfileMapper.map(userProfile.getPerson())).build();
-    }
-
-    @Override
-    public Response updateUserPreference(String id, String body) {
-        if (StringUtils.isEmpty(body)) {
-            throw new ConstraintException("Cannot update user preference with empty body",
-                    V1ErrorKeys.PREFERENCE_WITH_EMPTY_VALUE,
-                    null);
-        }
-        var userId = ApplicationContext.get().getPrincipal();
-
-        var preference = preferenceDAO.findById(id);
-        if (preference == null) {
-            return Response.status(Response.Status.NOT_FOUND).build();
-        }
-        if (!preference.getUserProfile().getUserId().equals(userId)) {
-            throw new ConstraintException("User does not match preference user id", V1ErrorKeys.PREFERENCE_NOT_FROM_ACTUAL_USER,
-                    null);
-        }
-        preference.setValue(body);
-        preference = preferenceDAO.update(preference);
-
-        return Response.ok(preferenceMapper.map(preference)).build();
-    }
-
-    @Override
-    public Response updateUserSettings(UpdateUserSettingsDTO updateUserSettingsDTO) {
-        var userId = ApplicationContext.get().getPrincipal();
-
-        var userProfile = userProfileDAO.getUserProfileByUserId(userId, UserProfile.ENTITY_GRAPH_LOAD_PERSON);
-        if (userProfile == null) {
-            return Response.status(Response.Status.NOT_FOUND).build();
-        }
-
-        userProfileMapper.update(userProfile.getAccountSettings(), updateUserSettingsDTO);
-        userProfile = userProfileDAO.update(userProfile);
-
-        return Response.ok(userProfileMapper.map(userProfile.getAccountSettings())).build();
-    }
-
-    @ServerExceptionMapper
-    public RestResponse<ProblemDetailResponseDTO> exception(ConstraintException ex) {
-        return exceptionMapper.exception(ex);
-    }
-
-    @ServerExceptionMapper
-    public RestResponse<ProblemDetailResponseDTO> constraint(ConstraintViolationException ex) {
-        return exceptionMapper.constraint(ex);
-    }
-
-    enum V1ErrorKeys {
-        USER_PROFILE_DOES_NOT_EXIST,
-
-        PREFERENCE_NOT_FROM_ACTUAL_USER,
-
-        PREFERENCE_WITH_EMPTY_VALUE,
-
-    }
 }
