+++ src/main/java/io/github/onecx/user/profile/rs/external/v1/mappers/PreferenceV1Mapper.java
@@ -6,25 +6,12 @@ import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.tkit.quarkus.rs.mappers.OffsetDateTimeMapper;
 
-import gen.io.github.onecx.user.profile.rs.external.v1.model.CreateUserPreferenceDTO;
 import gen.io.github.onecx.user.profile.rs.external.v1.model.UserPreferenceDTO;
 import gen.io.github.onecx.user.profile.rs.external.v1.model.UserPreferencesDTO;
 import io.github.onecx.user.profile.domain.models.Preference;
 
 @Mapper(uses = { OffsetDateTimeMapper.class })
-public interface PreferenceMapper {
-
-    @Mapping(target = "userProfile", ignore = true)
-    @Mapping(target = "tenantId", ignore = true)
-    @Mapping(target = "persisted", ignore = true)
-    @Mapping(target = "modificationUser", ignore = true)
-    @Mapping(target = "modificationDate", ignore = true)
-    @Mapping(target = "modificationCount", ignore = true)
-    @Mapping(target = "creationUser", ignore = true)
-    @Mapping(target = "creationDate", ignore = true)
-    @Mapping(target = "controlTraceabilityManual", ignore = true)
-    @Mapping(target = "id", ignore = true)
-    Preference create(CreateUserPreferenceDTO preferencesDTOv1s);
+public interface PreferenceV1Mapper {
 
     UserPreferenceDTO map(Preference preferences);
