+++ src/main/java/io/github/onecx/user/profile/rs/internal/mappers/AvatarMapper.java
@@ -1,14 +1,14 @@
-package io.github.onecx.user.profile.rs.external.v1.mappers;
+package io.github.onecx.user.profile.rs.internal.mappers;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.tkit.quarkus.rs.mappers.OffsetDateTimeMapper;
 
-import gen.io.github.onecx.user.profile.rs.external.v1.model.ImageInfoDTO;
+import gen.io.github.onecx.user.profile.rs.internal.model.ImageInfoDTO;
 import io.github.onecx.user.profile.domain.models.Image;
 
 @Mapper(uses = { OffsetDateTimeMapper.class })
-public interface AvatarV1Mapper {
+public interface AvatarMapper {
 
     @Mapping(target = "userUploaded", ignore = true)
     @Mapping(target = "smallImageUrl", ignore = true)
