+++ src/main/openapi/onecx-userprofile-v1-openapi.yaml
@@ -28,20 +28,6 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/ProblemDetailResponse'
-    delete:
-      tags:
-        - userProfileV1
-      description: Delete user profile for current user
-      operationId: deleteUserProfile
-      responses:
-        "204":
-          description: OK
-        "400":
-          description: Internal Server Error
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/ProblemDetailResponse'
   /v1/userProfile/me/preferences:
     get:
       tags:
@@ -61,81 +47,6 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/ProblemDetailResponse'
-    post:
-      tags:
-        - userProfileV1
-      description: Create user preference
-      operationId: createUserPreference
-      requestBody:
-        required: true
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/CreateUserPreference'
-      responses:
-        "201":
-          description: OK
-          headers:
-            Location:
-              required: true
-              schema:
-                type: string
-                format: url
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/UserPreference'
-        "400":
-          description: Bad request
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/ProblemDetailResponse'
-  /v1/userProfile/me/preferences/{id}:
-    patch:
-      tags:
-        - userProfileV1
-      description: Update preference value
-      operationId: updateUserPreference
-      parameters:
-        - $ref: '#/components/parameters/id'
-      requestBody:
-        required: true
-        content:
-          application/json:
-            schema:
-              type: string
-      responses:
-        "200":
-          description: OK
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/UserPreference'
-        "400":
-          description: Bad request
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/ProblemDetailResponse'
-        "404":
-          description: User preference not found
-    delete:
-      tags:
-        - userProfileV1
-      description: Delete preference by id for current user
-      operationId: deleteUserPreference
-      parameters:
-        - $ref: '#/components/parameters/id'
-      responses:
-        "204":
-          description: OK
-        "400":
-          description: Internal Server Error
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/ProblemDetailResponse'
   /v1/userProfile/me/person:
     get:
       tags:
@@ -145,30 +56,6 @@ paths:
       responses:
         "200":
           description: User person of current user
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/UserPerson'
-        "400":
-          description: Bad request
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/ProblemDetailResponse'
-    put:
-      tags:
-        - userProfileV1
-      description: Update person information
-      operationId: updateUserPerson
-      requestBody:
-        required: true
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/UpdateUserPerson'
-      responses:
-        "200":
-          description: OK
           content:
             application/json:
               schema:
@@ -198,30 +85,6 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/ProblemDetailResponse'
-    put:
-      tags:
-        - userProfileV1
-      description: Update user account settings
-      operationId: updateUserSettings
-      requestBody:
-        required: true
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/UpdateUserSettings'
-      responses:
-        "200":
-          description: OK
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/UserProfileAccountSettings'
-        "400":
-          description: Bad request
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/ProblemDetailResponse'
   /v1/userProfile/me/avatar:
     get:
       tags:
@@ -237,45 +100,6 @@ paths:
                 $ref: '#/components/schemas/ImageInfo'
         "404":
           description: User avatar not found
-    put:
-      tags:
-        - avatarV1
-      description: Upload user avatar
-      operationId: uploadAvatar
-      requestBody:
-        required: true
-        content:
-          'image/*':
-            schema:
-              type: string
-              format: binary
-      responses:
-        '200':
-          description: OK
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/ImageInfo'
-        "400":
-          description: Bad request
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/ProblemDetailResponse'
-    delete:
-      tags:
-        - avatarV1
-      description: Delete user's avatar
-      operationId: deleteUserAvatar
-      responses:
-        '204':
-          description: No Content
-        "400":
-          description: Bad request
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/ProblemDetailResponse'
   /v1/userProfile/me/avatar/{id}:
     get:
       tags:
@@ -302,17 +126,6 @@ paths:
           description: Not found
 components:
   schemas:
-    CreateUserPreference:
-      type: object
-      properties:
-        applicationId:
-          type: string
-        name:
-          type: string
-        description:
-          type: string
-        value:
-          type: string
     UserPreferences:
       type: object
       properties:
@@ -336,17 +149,6 @@ components:
       type: object
       properties:
         id:
-          type: string
-        version:
-          format: int32
-          type: integer
-        creationDate:
-          $ref: '#/components/schemas/OffsetDateTime'
-        creationUser:
-          type: string
-        modificationDate:
-          $ref: '#/components/schemas/OffsetDateTime'
-        modificationUser:
           type: string
         userId:
           type: string
@@ -415,19 +217,6 @@ components:
           $ref: '#/components/schemas/UserPersonAddress'
         phone:
           $ref: '#/components/schemas/UserPersonPhone'
-    UpdateUserSettings:
-      type: object
-      properties:
-        hideMyProfile:
-          type: boolean
-        locale:
-          type: string
-        timezone:
-          type: string
-        menuMode:
-          $ref: '#/components/schemas/MenuMode'
-        colorScheme:
-          $ref: '#/components/schemas/ColorScheme'
     UserProfileAccountSettings:
       type: object
       properties:
@@ -458,17 +247,6 @@ components:
     ImageInfo:
       type: object
       properties:
-        version:
-          format: int32
-          type: integer
-        creationDate:
-          $ref: '#/components/schemas/OffsetDateTime'
-        creationUser:
-          type: string
-        modificationDate:
-          $ref: '#/components/schemas/OffsetDateTime'
-        modificationUser:
-          type: string
         id:
           type: string
         userUploaded:
