+++ src/main/resources/application.properties
@@ -6,6 +6,7 @@ quarkus.datasource.jdbc.min-size=10
 quarkus.banner.enabled=false
 quarkus.hibernate-orm.database.generation=validate
 quarkus.hibernate-orm.multitenant=DISCRIMINATOR
+quarkus.hibernate-orm.jdbc.timezone=UTC
 quarkus.liquibase.migrate-at-start=true
 quarkus.liquibase.validate-on-migrate=true
 tkit.log.json.enabled=true
