+++ src/test/java/io/github/onecx/user/profile/rs/external/v1/controllers/AvatarV1RestControllerTest.java
@@ -13,7 +13,6 @@ import org.junit.jupiter.api.Test;
 import org.tkit.quarkus.test.WithDBData;
 
 import gen.io.github.onecx.user.profile.rs.external.v1.model.ImageInfoDTO;
-import gen.io.github.onecx.user.profile.rs.internal.model.ProblemDetailResponseDTO;
 import io.github.onecx.user.profile.test.AbstractTest;
 import io.quarkus.test.common.http.TestHTTPEndpoint;
 import io.quarkus.test.junit.QuarkusTest;
@@ -25,55 +24,34 @@ class AvatarV1RestControllerTest extends AbstractTest {
 
     @Test
     void testAvatarRestControler() throws URISyntaxException, IOException {
-        // test empty jpg image
-        File emptyAvatar = new File("src/test/resources/data/avatar_empty.jpg");
-        var error = given()
-                .when()
-                .contentType("image/jpg")
-                .body(emptyAvatar)
-                .header(APM_HEADER_PARAM, createToken("user1", null))
-                .put()
-                .then()
-                .statusCode(BAD_REQUEST.getStatusCode())
-                .extract().as(ProblemDetailResponseDTO.class);
-
-        assertThat(error).isNotNull();
-        assertThat(error.getErrorCode()).isEqualTo("UPLOAD_ERROR");
-
-        var avatarInfo = given()
+        given()
                 .when()
                 .contentType(APPLICATION_JSON)
-                .header(APM_HEADER_PARAM, createToken("user1", null))
+                .header(APM_HEADER_PARAM, createToken("user3", "org2"))
                 .get()
                 .then()
-                .statusCode(OK.getStatusCode())
-                .extract().as(ImageInfoDTO.class);
-
-        assertThat(avatarInfo).isNotNull();
-        assertThat(avatarInfo.getImageUrl()).isNotNull();
-        assertThat(avatarInfo.getSmallImageUrl()).isNotNull();
+                .statusCode(NOT_FOUND.getStatusCode());
 
         // add avatar
         File avatar = new File("src/test/resources/data/avatar_test.jpg");
 
-        var imageInfo = given()
+        // good tenant
+        var internalImageInfo = given().basePath("/internal/userProfile/me/avatar")
                 .when()
                 .contentType("image/jpg")
                 .body(avatar)
-                .header(APM_HEADER_PARAM, createToken("user1", null))
+                .header(APM_HEADER_PARAM, createToken("user1", "org1"))
                 .put()
                 .then()
                 .statusCode(OK.getStatusCode())
-                .extract().as(ImageInfoDTO.class);
-
-        assertThat(imageInfo).isNotNull();
-        assertThat(imageInfo.getSmallImageUrl()).isNotNull();
-        assertThat(imageInfo.getImageUrl()).isNotNull();
+                .extract().as(gen.io.github.onecx.user.profile.rs.internal.model.ImageInfoDTO.class);
 
         var smallAvatarByteArray = given()
                 .when()
-                .pathParam("id", imageInfo.getSmallImageUrl().substring(imageInfo.getSmallImageUrl().lastIndexOf("/") + 1))
-                .header(APM_HEADER_PARAM, createToken("user1", null))
+                .pathParam("id",
+                        internalImageInfo.getSmallImageUrl()
+                                .substring(internalImageInfo.getSmallImageUrl().lastIndexOf("/") + 1))
+                .header(APM_HEADER_PARAM, createToken("user1", "org1"))
                 .get("{id}")
                 .then()
                 .contentType("image/png")
@@ -84,8 +62,9 @@ class AvatarV1RestControllerTest extends AbstractTest {
 
         var avatarByteArray = given()
                 .when()
-                .pathParam("id", imageInfo.getImageUrl().substring(imageInfo.getImageUrl().lastIndexOf("/") + 1))
-                .header(APM_HEADER_PARAM, createToken("user1", null))
+                .pathParam("id",
+                        internalImageInfo.getImageUrl().substring(internalImageInfo.getImageUrl().lastIndexOf("/") + 1))
+                .header(APM_HEADER_PARAM, createToken("user1", "org1"))
                 .get("{id}")
                 .then()
                 .contentType("image/jpeg")
@@ -94,47 +73,25 @@ class AvatarV1RestControllerTest extends AbstractTest {
 
         assertThat(avatarByteArray).isNotNull();
 
-        avatarInfo = given()
+        var avatarInfo = given()
                 .when()
                 .contentType(APPLICATION_JSON)
-                .header(APM_HEADER_PARAM, createToken("user1", null))
+                .header(APM_HEADER_PARAM, createToken("user1", "org1"))
                 .get()
                 .then()
                 .statusCode(OK.getStatusCode())
                 .extract().as(ImageInfoDTO.class);
 
         assertThat(avatarInfo).isNotNull();
-        assertThat(avatarInfo.getImageUrl()).isNotNull().isEqualTo(imageInfo.getImageUrl());
-        assertThat(avatarInfo.getSmallImageUrl()).isNotNull().isEqualTo(imageInfo.getSmallImageUrl());
-
-        given()
-                .when()
-                .contentType(APPLICATION_JSON)
-                .header(APM_HEADER_PARAM, createToken("user1", null))
-                .delete()
-                .then()
-                .statusCode(NO_CONTENT.getStatusCode());
-
-        given()
-                .when()
-                .contentType(APPLICATION_JSON)
-                .header(APM_HEADER_PARAM, createToken("user1", null))
-                .get()
-                .then()
-                .statusCode(NOT_FOUND.getStatusCode());
+        assertThat(avatarInfo.getImageUrl().substring(avatarInfo.getImageUrl().indexOf("userProfile"))).isNotNull()
+                .isEqualTo(internalImageInfo.getImageUrl().substring(internalImageInfo.getImageUrl().indexOf("userProfile")));
+        assertThat(avatarInfo.getSmallImageUrl().substring(avatarInfo.getSmallImageUrl().indexOf("userProfile"))).isNotNull()
+                .isEqualTo(internalImageInfo.getSmallImageUrl()
+                        .substring(internalImageInfo.getSmallImageUrl().indexOf("userProfile")));
     }
 
     @Test
     void testAvatarRestControlerErrorStates() {
-        // delete from not existing profile
-        given()
-                .when()
-                .contentType(APPLICATION_JSON)
-                .header(APM_HEADER_PARAM, createToken("not-existing", null))
-                .delete()
-                .then()
-                .statusCode(NO_CONTENT.getStatusCode());
-
         // get not existing image
         given()
                 .when()
