+++ src/test/java/io/github/onecx/user/profile/rs/internal/controllers/AvatarRestControllerIT.java
@@ -3,5 +3,5 @@ package io.github.onecx.user.profile.rs.internal.controllers;
 import io.quarkus.test.junit.QuarkusIntegrationTest;
 
 @QuarkusIntegrationTest
-public class UserProfileInternalRestControllerTenantIT extends UserProfileInternalRestControllerTenantTest {
+public class AvatarRestControllerIT extends AvatarRestControllerTest {
 }
