+++ src/test/java/io/github/onecx/user/profile/rs/internal/controllers/AvatarRestControllerTenantTest.java
@@ -1,4 +1,4 @@
-package io.github.onecx.user.profile.rs.external.v1.controllers;
+package io.github.onecx.user.profile.rs.internal.controllers;
 
 import static io.restassured.RestAssured.given;
 import static jakarta.ws.rs.core.MediaType.APPLICATION_JSON;
@@ -12,34 +12,33 @@ import java.net.URISyntaxException;
 import org.junit.jupiter.api.Test;
 import org.tkit.quarkus.test.WithDBData;
 
-import gen.io.github.onecx.user.profile.rs.external.v1.model.ImageInfoDTO;
+import gen.io.github.onecx.user.profile.rs.internal.model.*;
 import io.github.onecx.user.profile.test.AbstractTest;
 import io.quarkus.test.common.http.TestHTTPEndpoint;
 import io.quarkus.test.junit.QuarkusTest;
 
 @QuarkusTest
-@TestHTTPEndpoint(AvatarV1RestController.class)
+@TestHTTPEndpoint(AvatarRestController.class)
 @WithDBData(value = "data/testdata.xml", deleteBeforeInsert = true, deleteAfterTest = true, rinseAndRepeat = true)
-class AvatarV1RestControllerTenantTest extends AbstractTest {
+class AvatarRestControllerTenantTest extends AbstractTest {
 
     @Test
     void testAvatarRestControler() throws URISyntaxException, IOException {
-        var testInfoDto = given()
-                .when()
-                .contentType(APPLICATION_JSON)
-                .header(APM_HEADER_PARAM, createToken("user1", "org1"))
-                .get()
-                .then()
-                .statusCode(OK.getStatusCode())
-                .extract().as(ImageInfoDTO.class);
-
-        assertThat(testInfoDto.getUserUploaded()).isNull();
-
         // add avatar
         File avatar = new File("src/test/resources/data/avatar_test.jpg");
 
+        // wrong tenant
+        given()
+                .when()
+                .contentType("image/jpg")
+                .body(avatar)
+                .header(APM_HEADER_PARAM, createToken("user1", "org2"))
+                .put()
+                .then()
+                .statusCode(NOT_FOUND.getStatusCode());
+
         // good tenant
-        var internalImageInfo = given().basePath("/internal/userProfile/me/avatar")
+        var imageInfo = given()
                 .when()
                 .contentType("image/jpg")
                 .body(avatar)
@@ -47,14 +46,12 @@ class AvatarV1RestControllerTenantTest extends AbstractTest {
                 .put()
                 .then()
                 .statusCode(OK.getStatusCode())
-                .extract().as(gen.io.github.onecx.user.profile.rs.internal.model.ImageInfoDTO.class);
+                .extract().as(ImageInfoDTO.class);
 
         // wrong tenant get small avatar
         given()
                 .when()
-                .pathParam("id",
-                        internalImageInfo.getSmallImageUrl()
-                                .substring(internalImageInfo.getSmallImageUrl().lastIndexOf("/") + 1))
+                .pathParam("id", imageInfo.getSmallImageUrl().substring(imageInfo.getSmallImageUrl().lastIndexOf("/") + 1))
                 .header(APM_HEADER_PARAM, createToken("user1", "org3"))
                 .get("{id}")
                 .then()
@@ -63,9 +60,7 @@ class AvatarV1RestControllerTenantTest extends AbstractTest {
         // good tenant get small avatar
         var smallAvatarByteArray = given()
                 .when()
-                .pathParam("id",
-                        internalImageInfo.getSmallImageUrl()
-                                .substring(internalImageInfo.getSmallImageUrl().lastIndexOf("/") + 1))
+                .pathParam("id", imageInfo.getSmallImageUrl().substring(imageInfo.getSmallImageUrl().lastIndexOf("/") + 1))
                 .header(APM_HEADER_PARAM, createToken("user1", "org1"))
                 .get("{id}")
                 .then()
@@ -95,10 +90,41 @@ class AvatarV1RestControllerTenantTest extends AbstractTest {
                 .extract().as(ImageInfoDTO.class);
 
         assertThat(avatarInfo).isNotNull();
-        assertThat(avatarInfo.getImageUrl().substring(avatarInfo.getImageUrl().indexOf("userProfile"))).isNotNull()
-                .isEqualTo(internalImageInfo.getImageUrl().substring(internalImageInfo.getImageUrl().indexOf("userProfile")));
-        assertThat(avatarInfo.getSmallImageUrl().substring(avatarInfo.getSmallImageUrl().indexOf("userProfile"))).isNotNull()
-                .isEqualTo(internalImageInfo.getSmallImageUrl()
-                        .substring(internalImageInfo.getSmallImageUrl().indexOf("userProfile")));
+        assertThat(avatarInfo.getImageUrl()).isNotNull().isEqualTo(imageInfo.getImageUrl());
+        assertThat(avatarInfo.getSmallImageUrl()).isNotNull().isEqualTo(imageInfo.getSmallImageUrl());
+
+        // delete with wrong tenant
+        given()
+                .when()
+                .contentType(APPLICATION_JSON)
+                .header(APM_HEADER_PARAM, createToken("user1", "org2"))
+                .delete()
+                .then()
+                .statusCode(NO_CONTENT.getStatusCode());
+
+        given()
+                .when()
+                .contentType(APPLICATION_JSON)
+                .header(APM_HEADER_PARAM, createToken("user1", "org1"))
+                .get()
+                .then()
+                .statusCode(OK.getStatusCode());
+
+        // delete with good tenant
+        given()
+                .when()
+                .contentType(APPLICATION_JSON)
+                .header(APM_HEADER_PARAM, createToken("user1", "org1"))
+                .delete()
+                .then()
+                .statusCode(NO_CONTENT.getStatusCode());
+
+        given()
+                .when()
+                .contentType(APPLICATION_JSON)
+                .header(APM_HEADER_PARAM, createToken("user1", "org1"))
+                .get()
+                .then()
+                .statusCode(NOT_FOUND.getStatusCode());
     }
 }
