+++ src/test/java/io/github/onecx/user/profile/rs/internal/controllers/UserProfileAdminRestControllerIT.java
@@ -1,0 +1,7 @@
+package io.github.onecx.user.profile.rs.internal.controllers;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class UserProfileAdminRestControllerIT extends UserProfileAdminRestControllerTest {
+}
