+++ src/test/java/io/github/onecx/user/profile/rs/internal/controllers/UserProfileAdminRestControllerTenantIT.java
@@ -1,0 +1,7 @@
+package io.github.onecx.user.profile.rs.internal.controllers;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class UserProfileAdminRestControllerTenantIT extends UserProfileAdminRestControllerTenantTest {
+}
