+++ src/test/java/io/github/onecx/user/profile/rs/internal/controllers/UserProfileAdminRestControllerTest.java
@@ -9,17 +9,31 @@ import org.junit.jupiter.api.Test;
 import org.tkit.quarkus.test.WithDBData;
 
 import gen.io.github.onecx.user.profile.rs.internal.model.*;
+import io.github.onecx.user.profile.rs.internal.mappers.InternalExceptionMapper;
 import io.github.onecx.user.profile.test.AbstractTest;
 import io.quarkus.test.common.http.TestHTTPEndpoint;
 import io.quarkus.test.junit.QuarkusTest;
 
 @QuarkusTest
-@TestHTTPEndpoint(UserProfileInternalRestController.class)
+@TestHTTPEndpoint(UserProfileAdminRestController.class)
 @WithDBData(value = "data/testdata.xml", deleteBeforeInsert = true, deleteAfterTest = true, rinseAndRepeat = true)
-class UserProfileInternalRestControllerTenantTest extends AbstractTest {
+class UserProfileAdminRestControllerTest extends AbstractTest {
 
     @Test
     void createUserProfileTest() {
+        // test not sending the body
+        var error = given()
+                .when()
+                .contentType(APPLICATION_JSON)
+                .post()
+                .then()
+                .statusCode(BAD_REQUEST.getStatusCode())
+                .extract().as(ProblemDetailResponseDTO.class);
+
+        assertThat(error).isNotNull();
+        assertThat(error.getErrorCode()).isEqualTo(InternalExceptionMapper.TechnicalErrorKeys.CONSTRAINT_VIOLATIONS.name());
+        assertThat(error.getDetail()).isEqualTo("createUserProfile.createUserProfileRequestDTO: must not be null");
+
         // create user profile with content
         CreateUserProfileRequestDTO request = new CreateUserProfileRequestDTO();
         request.setUserId("cap");
@@ -36,7 +50,7 @@ class UserProfileInternalRestControllerTenantTest extends AbstractTest {
         var result = given()
                 .when()
                 .contentType(APPLICATION_JSON)
-                .header(APM_HEADER_PARAM, createToken("user4", "org3"))
+                .header(APM_HEADER_PARAM, createToken("user3", "org2"))
                 .body(request)
                 .post()
                 .then()
@@ -48,18 +62,10 @@ class UserProfileInternalRestControllerTenantTest extends AbstractTest {
         assertThat(result.getModificationDate()).isNotNull();
         assertThat(result.getPerson().getPhone()).isNull();
 
-        given()
-                .when()
-                .pathParam("id", "cap")
-                .header(APM_HEADER_PARAM, createToken("user4", "org2"))
-                .get("{id}")
-                .then()
-                .statusCode(NOT_FOUND.getStatusCode());
-
         result = given()
                 .when()
+                .header(APM_HEADER_PARAM, createToken("user3", "org2"))
                 .pathParam("id", "cap")
-                .header(APM_HEADER_PARAM, createToken("user4", "org3"))
                 .get("{id}")
                 .then()
                 .statusCode(OK.getStatusCode())
@@ -70,41 +76,43 @@ class UserProfileInternalRestControllerTenantTest extends AbstractTest {
         assertThat(result.getModificationDate()).isNotNull();
         assertThat(result.getPerson().getPhone()).isNull();
 
+        error = given()
+                .when()
+                .contentType(APPLICATION_JSON)
+                .header(APM_HEADER_PARAM, createToken("user3", "org2"))
+                .body(request)
+                .post()
+                .then()
+                .statusCode(BAD_REQUEST.getStatusCode())
+                .extract().as(ProblemDetailResponseDTO.class);
+
+        assertThat(error).isNotNull();
+        assertThat(error.getErrorCode()).isEqualTo("PERSIST_ENTITY_FAILED");
+        assertThat(error.getDetail()).isEqualTo(
+                "could not execute statement [ERROR: duplicate key value violates unique constraint 'ukm9nl9w7ih2pti88rq0xf31c5y'  Detail: Key (user_id, tenant_id)=(cap, tenant-200) already exists.]");
     }
 
     @Test
     void deleteUserProfileTest() {
-        // delete existing profile with wrong tenant
+        // delete not existing profile
         given()
                 .when()
-                .header(APM_HEADER_PARAM, createToken("user4", "org3"))
+                .pathParam("id", "not-existing")
+                .delete("{id}")
+                .then()
+                .statusCode(NO_CONTENT.getStatusCode());
+
+        // delete existing profile
+        given()
+                .when()
                 .pathParam("id", "user1")
                 .delete("{id}")
                 .then()
                 .statusCode(NO_CONTENT.getStatusCode());
 
-        // load deleted profile still found as deleted with wrong tenant
+        // load deleted profile
         given()
                 .when()
-                .header(APM_HEADER_PARAM, createToken("user4", "org1"))
-                .pathParam("id", "user1")
-                .get("{id}")
-                .then()
-                .statusCode(OK.getStatusCode());
-
-        // delete existing profile with correct tenant
-        given()
-                .when()
-                .header(APM_HEADER_PARAM, createToken("user4", "org1"))
-                .pathParam("id", "user1")
-                .delete("{id}")
-                .then()
-                .statusCode(NO_CONTENT.getStatusCode());
-
-        // load deleted profile returns not found
-        given()
-                .when()
-                .header(APM_HEADER_PARAM, createToken("user4", "org1"))
                 .pathParam("id", "user1")
                 .get("{id}")
                 .then()
@@ -113,19 +121,9 @@ class UserProfileInternalRestControllerTenantTest extends AbstractTest {
 
     @Test
     void getUserProfileTest() {
-        // get with different tenant
-        given()
-                .when()
-                .header(APM_HEADER_PARAM, createToken("user4", "org2"))
-                .pathParam("id", "user1")
-                .get("{id}")
-                .then()
-                .statusCode(NOT_FOUND.getStatusCode());
-
-        // get existing profile with correct tenant
+        // get existing profile
         var result = given()
                 .when()
-                .header(APM_HEADER_PARAM, createToken("user4", "org1"))
                 .pathParam("id", "user1")
                 .get("{id}")
                 .then()
@@ -141,13 +139,24 @@ class UserProfileInternalRestControllerTenantTest extends AbstractTest {
 
     @Test
     void searchUserProfileTest() {
+        // search without criteria
+        var error = given()
+                .when()
+                .contentType(APPLICATION_JSON)
+                .post("search")
+                .then()
+                .statusCode(BAD_REQUEST.getStatusCode())
+                .extract().as(ProblemDetailResponseDTO.class);
+
+        assertThat(error).isNotNull();
+        assertThat(error.getErrorCode()).isEqualTo(InternalExceptionMapper.TechnicalErrorKeys.CONSTRAINT_VIOLATIONS.name());
+        assertThat(error.getDetail()).isEqualTo("searchUserProfile.userPersonCriteriaDTO: must not be null");
+
         // search with criteria
-        // org1
         UserPersonCriteriaDTO criteriaDTO = new UserPersonCriteriaDTO();
         criteriaDTO.setEmail("*cap.de");
         var result = given()
                 .when()
-                .header(APM_HEADER_PARAM, createToken("user4", "org1"))
                 .body(criteriaDTO)
                 .contentType(APPLICATION_JSON)
                 .post("search")
@@ -160,10 +169,10 @@ class UserProfileInternalRestControllerTenantTest extends AbstractTest {
         var resultList = result.getStream();
         assertThat(resultList.get(0).getPerson().getLastName()).isEqualTo("One");
 
-        // org2
+        // search with empty criteria
+        criteriaDTO = new UserPersonCriteriaDTO();
         result = given()
                 .when()
-                .header(APM_HEADER_PARAM, createToken("user4", "org2"))
                 .body(criteriaDTO)
                 .contentType(APPLICATION_JSON)
                 .post("search")
@@ -172,17 +181,65 @@ class UserProfileInternalRestControllerTenantTest extends AbstractTest {
                 .extract().as(UserProfilePageResultDTO.class);
 
         assertThat(result).isNotNull();
-        assertThat(result.getTotalElements()).isEqualTo(1);
+        assertThat(result.getTotalElements()).isEqualTo(2);
         resultList = result.getStream();
-        assertThat(resultList.get(0).getPerson().getLastName()).isEqualTo("Three");
+        assertThat(resultList.get(0).getPerson().getLastName()).isEqualTo("One");
 
+        // search with all criteria filled
+        criteriaDTO = new UserPersonCriteriaDTO();
+        criteriaDTO.setEmail("email");
+        criteriaDTO.setUserId("userId");
+        criteriaDTO.setFirstName("firstName");
+        criteriaDTO.setLastName("lastName");
+        result = given()
+                .when()
+                .body(criteriaDTO)
+                .contentType(APPLICATION_JSON)
+                .post("search")
+                .then()
+                .statusCode(OK.getStatusCode())
+                .extract().as(UserProfilePageResultDTO.class);
+
+        assertThat(result).isNotNull();
+        assertThat(result.getTotalElements()).isZero();
+
+        // search with all criteria filled with * or ?
+        criteriaDTO = new UserPersonCriteriaDTO();
+        criteriaDTO.setEmail("*cap.d?");
+        criteriaDTO.setUserId("user?");
+        criteriaDTO.setFirstName("User*");
+        criteriaDTO.setLastName("*o*");
+        result = given()
+                .when()
+                .body(criteriaDTO)
+                .contentType(APPLICATION_JSON)
+                .post("search")
+                .then()
+                .statusCode(OK.getStatusCode())
+                .extract().as(UserProfilePageResultDTO.class);
+
+        assertThat(result).isNotNull();
+        assertThat(result.getTotalElements()).isEqualTo(2);
     }
 
     @Test
     void updateUserProfileTest() {
+        // no content for update
+        var error = given()
+                .when()
+                .contentType(APPLICATION_JSON)
+                .pathParam("id", "not-existing")
+                .put("{id}")
+                .then()
+                .statusCode(BAD_REQUEST.getStatusCode())
+                .extract().as(ProblemDetailResponseDTO.class);
+
+        assertThat(error).isNotNull();
+        assertThat(error.getErrorCode()).isEqualTo(InternalExceptionMapper.TechnicalErrorKeys.CONSTRAINT_VIOLATIONS.name());
+        assertThat(error.getDetail()).isEqualTo("updateUserProfile.updateUserPersonRequestDTO: must not be null");
+
         var userProfileDTO = given()
                 .when()
-                .header(APM_HEADER_PARAM, createToken("user4", "org1"))
                 .pathParam("id", "user1")
                 .get("{id}")
                 .then()
@@ -198,22 +255,21 @@ class UserProfileInternalRestControllerTenantTest extends AbstractTest {
         requestDTO.setPhone(userProfileDTO.getPerson().getPhone());
         requestDTO.getPhone().setNumber("123456789");
         requestDTO.getPhone().setType(PhoneTypeDTO.LANDLINE);
+        requestDTO.setModificationCount(userProfileDTO.getModificationCount());
 
-        // update existing profile with wrong tenant
+        // update not existing profile
         given()
                 .when()
-                .header(APM_HEADER_PARAM, createToken("user4", "org2"))
                 .contentType(APPLICATION_JSON)
                 .body(requestDTO)
-                .pathParam("id", "user1")
+                .pathParam("id", "not-existing")
                 .put("{id}")
                 .then()
                 .statusCode(NOT_FOUND.getStatusCode());
 
-        // update existing profile with correct tenant
+        // update existing profile
         given()
                 .when()
-                .header(APM_HEADER_PARAM, createToken("user4", "org1"))
                 .contentType(APPLICATION_JSON)
                 .body(requestDTO)
                 .pathParam("id", "user1")
@@ -221,9 +277,21 @@ class UserProfileInternalRestControllerTenantTest extends AbstractTest {
                 .then()
                 .statusCode(NO_CONTENT.getStatusCode());
 
+        // update 2nd time - existing profile - get optimistic lock exception
+        error = given()
+                .when()
+                .contentType(APPLICATION_JSON)
+                .body(requestDTO)
+                .pathParam("id", "user1")
+                .put("{id}")
+                .then()
+                .statusCode(BAD_REQUEST.getStatusCode())
+                .extract().as(ProblemDetailResponseDTO.class);
+
+        assertThat(error.getErrorCode()).isEqualTo(InternalExceptionMapper.TechnicalErrorKeys.OPTIMISTIC_LOCK.name());
+
         userProfileDTO = given()
                 .when()
-                .header(APM_HEADER_PARAM, createToken("user4", "org1"))
                 .pathParam("id", "user1")
                 .get("{id}")
                 .then()
