+++ src/test/java/io/github/onecx/user/profile/rs/internal/controllers/UserProfileRestControllerTenantIT.java
@@ -3,5 +3,5 @@ package io.github.onecx.user.profile.rs.internal.controllers;
 import io.quarkus.test.junit.QuarkusIntegrationTest;
 
 @QuarkusIntegrationTest
-public class UserProfileInternalRestControllerIT extends UserProfileInternalRestControllerTest {
+public class UserProfileRestControllerTenantIT extends UserProfileRestControllerTenantTest {
 }
